[request] #33
Hi, again.
I started learning a bit and got my efforts for the first request for an request ticket a while back
https://github.com/sualfred/script.embuary.helper/issues/33
So, may code is ugly or needs improvement.
But it's does'nt give any error in the logs.

I really hope you found it quiet usefull too.

Usecase if not self explaining:

If Select an item if custom container is filled with the playlistmusic:://
It'll play the selected item, but keeps the the list itself 'intact' ()
(kodi default select|playmedia() is play the selected song and clear the current list ,
 if using your helper metheod from here, it clears all previous items out of the list.

I kept your default predefinded param (music|video), to differ between the playlist ids, so should also work also for video content , though i didnt tested it myself.

Best wishes...
This is how it working for me.
https://streamable.com/m1s1qf